+++ packages/landing/src/web/Footer/Footer.stories.tsx
@@ -1,9 +1,11 @@
 import React from 'react'
-import Footer from './Footer'
 import { MOCK_FOOTER_PROPS } from '../../mocks'
+import getStorybookTitle from '../../utils/getStorybookTitle'
+import Footer from './Footer'
 
 export default {
   component: Footer,
+  title: getStorybookTitle(Footer.displayName),
   parameters: { layout: 'fullscreen' },
 }