+++ packages/ui/src/core/Box.stories.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
+import { getCoreStorybookTitle } from '../utils/getStorybookTitle'
 import Box from './Box'
 
 export default {
-  title: 'ui/Box',
+  title: getCoreStorybookTitle('Box'),
   component: Box,
   parameters: { layout: 'fullscreen' },
   args: {