+++ packages/ui/src/core/Header/SubHeader.stories.tsx
@@ -1,8 +1,10 @@
 import React from 'react'
-import SubHeader from './SubHeader'
 import { MOCK_SUB_HEADER_PROPS } from '../../mocks'
+import { getCoreStorybookTitle } from '../../utils/getStorybookTitle'
+import SubHeader from './SubHeader'
 
 export default {
+  title: getCoreStorybookTitle(SubHeader.name),
   component: SubHeader,
   parameters: { layout: 'fullscreen' },
   args: MOCK_SUB_HEADER_PROPS,